Update README to include instructions for building CLI only
Users may want to play around with gpt4all-j from the command line. But they may not have Qt, and might not want to get it, or may find it very hard to do so - eg when using a Google Colab or similar hosted service.

It's easy to build the CLI tools just by building the `ggml` sub folder.  So this commit adds instructions on doing that, including an example invocation of the `gpt-j` binary.