chatllm: use unique_ptr for LLModelInfo::model
This fixes a memory leak if there was a model in the store on exit, e.g.
if the user loads a model and then switches to a chat associated with a
different model file without loading it.

ChatLLM::destroyStore is added to fix a heap-use-after-free caused by
the unique_ptr being freed too late. Global destructors are hard.

Signed-off-by: Jared Van Bortel <jared@nomic.ai>