database: fix reading of UTF-16 and UTF-32 files
QTextStream is capable of automatically detecting these encodings if the
file starts with a BOM. However, this is only done if the first byte of
the stream is read. Seeking immediately after opening the stream
prevents this, so we need to read a byte first, or we will get garbage
data in later calls to scanQueue() on the same document.

Signed-off-by: Jared Van Bortel <jared@nomic.ai>