backend: fix CUDA fallback using wrong device and lying
When we do CPU fallback and request CUDA with ngl=0, it still uses the
GPU for matrix multipliciation. It shouldn't clear the GPU device on
failure, as it is still relevant when we fall back. We also shouldn't
allow loadModel with an unset device.

This fixes a case where gpuDeviceName() would return "", because
usingGPUDevice() reported true while we had no known device name.

Signed-off-by: Jared Van Bortel <jared@nomic.ai>