Models3.json Add Yi-1.5-9B-Chat-16K
Adds model support for [Yi-1.5-9B-Chat-16K](https://huggingface.co/GPT4All-Community/Yi-1.5-9B-Chat-16K-GGUF)

## Description:

It is a bilingual model and at the date of writing with strong results in benchmarks (for its parameter size). It supports a context of up to 16K.

- Minimum required version: GPT4All 3.1.
- The model was trained on English and Chinese language.
- License: Apache 2.0
- Q4_0

## Personal Impression:
I got the impression the model is very task focused and this is the reason, why I chose  `Below is an instruction that describes a task. Write a response that appropriately completes the request.` as system prompt. I have seen refusals when it was tasked with certain things and has the typical "know it better than the user" vibe and seems to be finetuned for being a professional assistant. For instance, roleplay caused refusals, but writing a cover letter was no problem. Its long context and quality of responses makes it a good model, if you can bear its alignment. It mainly will appeal to English and Chinese speaking users.

Signed-off-by: ThiloteE <73715071+ThiloteE@users.noreply.github.com>