llamamodel: remove n_last_batch_tokens
llama.cpp commit e3c337d87 ("llama : support negative ith in llama_get_
API (#6519)") added a simpler way to get the logits for the last token
in the batch, so use that instead. This also fixes potential issues with
not serializing this value with the rest of the prompt context, although
in practice we should always call evalTokens before
llama_sample_top_p_top_k.

Signed-off-by: Jared Van Bortel <jared@nomic.ai>