chatllm: explicitly serialize a state version for remote models
This fixes a Q_UNREACHABLE assertion when saving a chat with a remote
model in a debug build.

This should be a non-functional change for release builds, since the
compiler has been optimizing the Q_UNREACHABLE() to write a state
version for remote models too. But we shouldn't rely on this, because in
the next GCC version there could be nose demons instead.

Signed-off-by: Jared Van Bortel <jared@nomic.ai>