database: cache known document ids for fast lookup
Performing a SQL query over the chunks table every time we call
ChunkStreamer::setDocument with a new document is very slow, be it a
delete or a select or otherwise. To alleviate this, cache the list of
document ids that might need to be removed. This way, we can keep
acceptable performance while maintaining guarantees that there are no
duplicate chunks regardless of how ChunkStreamer is called.

Signed-off-by: Jared Van Bortel <jared@nomic.ai>