modellist: use blank default template and system prompt
Alpaca is obsolete. We would be much better off defaulting to the chat
template that is shipped with the model.

The default system prompt exists in order to demonstrate how it should
be templated with Alpaca, but this is no longer applicable because this
is now part of the Jinja template.

Signed-off-by: Jared Van Bortel <jared@nomic.ai>