Add olmoe and granitemoe architecture to whitelist
Those two model architectures are supported in llama.cpp and technically in GPT4All 3.8.0 as well, but they have to be whitelisted to make them accessible to users, so do it.

Here are the models:
- https://huggingface.co/allenai/OLMoE-1B-7B-0924-Instruct
- https://huggingface.co/ibm-granite/granite-3.1-3b-a800m-instruct

They are really small and not the brightest, but very fast. I am hoping to make use of them for sentence similarity search with the LocalDocs feature. There was also a user that wanted to have tiny LLMs for educational purposes in Africa, because of hardware constraints.

Signed-off-by: ThiloteE <73715071+ThiloteE@users.noreply.github.com>